+++ src/driver/mod.rs
@@ -62,7 +62,7 @@ pub(crate) use self::{
         AttachmentInfo, AttachmentRef, FramebufferAttachmentImageInfo, FramebufferInfo, RenderPass,
         RenderPassInfo, SubpassDependency, SubpassInfo,
     },
-    shader::{DescriptorBinding, DescriptorBindingMap, DescriptorInfo},
+    shader::{Descriptor, DescriptorBindingMap, DescriptorInfo},
     surface::Surface,
 };
 
@@ -413,7 +413,7 @@ fn merge_push_constant_ranges(pcr: &[vk::PushConstantRange]) -> Vec<vk::PushCons
 
                 if lhs.offset == rhs.offset && lhs.size == rhs.size {
                     res[i].stage_flags |= rhs.stage_flags;
-                    res.remove(j);
+                    let _ = res.remove(j);
                 } else if lhs.offset == rhs.offset {
                     res[i].stage_flags |= rhs.stage_flags;
                     res[j].offset += lhs.size;