+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Updated `ash` to v0.38
+- Updated `winit` to v0.30 (_and moved related functionality to new `screen-13-window` crate_)
+
+### Removed
+
+- `log` and `winit` are no longer exported by `use screen_13::prelude::*`
+
 ## [0.11.3] - 2024-05-29
 
 ### Added